overcomplicated mumbo jumbo
player movement is more clearly defined. taking a data oriented approach hopefully. camera class is over defined, just use explicit target definition on player

+++ camera.h
@@ -15,6 +15,6 @@ public:
 	void set_speed();
 
 	Camera2D player_camera;
-	float camera_speed;
+	int camera_speed;
 
 };
